requirements.txt
django==4.1.2
django-tempus-dominus==5.39.0